internal/evaluator/documentation/documentation_test.go
// Copyright The Enterprise Contract Contributors
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//
// SPDX-License-Identifier: Apache-2.0

//go:build unit

package main

import (
	"io/fs"
	"os"
	"path/filepath"
	"testing"

	"github.com/gkampitakis/go-snaps/snaps"
	"github.com/stretchr/testify/require"
)

func TestWriteBuiltinsToYAML(t *testing.T) {
	dir := t.TempDir()
	err := writeBultinsToYAML(dir)
	require.NoError(t, err)
	err = filepath.WalkDir(dir, func(path string, d fs.DirEntry, err error) error {
		require.NoError(t, err)
		if d.IsDir() {
			return nil
		}
		contents, err := os.ReadFile(path)
		require.NoError(t, err)
		snaps.MatchSnapshot(t, string(contents))
		return nil
	})
	require.NoError(t, err)
}
